TRegex: refactored DFANodeSplit

+++ graal-js/src/com.oracle.truffle.regex/src/com/oracle/truffle/regex/tregex/automaton/IndexedState.java
@@ -5,5 +5,5 @@
 package com.oracle.truffle.regex.tregex.automaton;
 
 public interface IndexedState {
-    int getId();
+    short getId();
 }
